"AI" / get bot to talk / chat lockdown / more
1) Added an AI class to keep track of a bunch of triggers and responses
so you can communicate with it in english

2) Added onPrivateMessage so you can now get the chat to go +i if its
removed and for users to pm the bot with a certain string for it to
invite you into the chat

3) Added code in onPrivateMessage to talk as the bot in chat

4) changed the YouTube link reader. It should grab all youtube links no
matter what now (except if there is other text sent with the link.
Havent quite gotten that far)

5) Added a submitter field in Links class so we now know who submitted
such glorious links to the bot (also making it able to, in the future,
add some type of reward/xp type deal to make the chat a bit more fun)

6) Added 2 more fields in the config pertaining specifically to the
invite only settings.

7) A few little things i probably have fiddled with here and there (been
working on it on and off since last commit